requirements.txt
# Core ML Framework
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Tokenizers
tokenizers>=0.15.0
sentencepiece>=0.1.99
transformers>=4.36.0

# Training Optimization
deepspeed>=0.12.0
accelerate>=0.25.0
flash-attn>=2.5.0

# Data Processing
datasets>=2.16.0
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0

# Language Detection
lingua-language-detector>=2.0.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Monitoring & Logging
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Utilities
rich>=13.7.0
click>=8.1.0
python-dotenv>=1.0.0

# Remote Dataset Support
boto3>=1.34.0
requests>=2.31.0
huggingface-hub>=0.20.0

# Analysis & Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
